pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gyre"
authors = [{ name = "Hamish Friedlander", email = "hafriedlander@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
dependencies = [
  # Core pipeline
  "torch ~= 1.12.1",
  "einops ~= 0.5.0",
  "torchvision ~= 0.13.1",
  "numpy ~= 1.23.3",
  "opencv-python-headless ~= 4.6.0.66",
  "scipy ~= 1.9.1",
  "ftfy ~= 6.1.1",
  "transformers ~= 4.26.1",
  "diffusers ~= 0.14.0",
  "accelerate ~= 0.16.0",
  "easing-functions ~= 1.0.4",
  # For ToMe
  "timm ~= 0.6.11",
  # For Structured Diffusion
  "nltk ~= 3.7",
  "stanza ~= 1.4.2",
  # For K-Diffusion
  "torchdiffeq ~= 0.2.3",
  "torchsde ~= 0.2.5",
  # For MiDaS
  "imutils ~= 0.5.4",
  # For Hinters
  "kornia ~= 0.6.10",
  # For mmXXX
  "mmcv-full ~= 1.7.1",
  "json-tricks ~= 3.16.1",
  "prettytable ~= 3.6.0",
  "terminaltables ~= 3.1.10",
  "munkres ~= 1.1.4",
  # For Server
  "protobuf ~= 3.20",
  "grpcio ~= 1.48.1",
  "wsgicors ~= 0.7.0",
  "Twisted ~= 22.8.0",
  "hupper ~= 1.10.3",
  "watchdog ~= 2.1.9",
  "python-dotenv ~= 0.21.0",
  "service_identity ~= 21.1.0",
  "safetensors ~= 0.2.6",
  "multipart ~= 0.2.4",
  "accept-types ~= 0.4.1",
  "gdown ~= 4.6.2",
  "colorama ~= 0.4.6",
  # For client
  "py-machineid ~= 0.3.0",
  # For Tests
  "pynvml ~= 11.4.1",
]

[project.optional-dependencies]
# These are used by the VSCode configuration to format the python code consistently
dev = [
  "black ~= 22.10.0", 
  "flake8 ~= 6.0.0", 
  "flake8-pyproject ~= 1.2.1"
]
# This is used by the Dockerfile as a minimal install for building mmcv
core = [
  "torch ~= 1.12.1",
  "einops ~= 0.5.0",
  "torchvision ~= 0.13.1",
  "numpy ~= 1.23.3",
  "opencv-python-headless ~= 4.6.0.66",
  "scipy ~= 1.9.1",
]

[project.urls]
Home = "https://github.com/stablecabal/gyre"

[project.scripts]
gyre = "gyre.server:main"

[tool.flit.module]
name = "gyre"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
select = "C,E,F,W,B,B950"
extend-ignore = "E203, E501, W503"
